+++ src/rules/for_with_begin.rs
@@ -38,6 +38,10 @@ impl Rule for ForWithBegin {
     }
 
     fn hint(&self) -> String {
-        String::from("multiline 'for' statement must have 'begin'")
+        String::from("multiline `for` statement must have `begin`")
+    }
+
+    fn reason(&self) -> String {
+        String::from("if there is not `begin`, the second statatement are confusing")
     }
 }